Extract "equals" assertion to own file
